+++ NEWS.md
@@ -3,6 +3,10 @@
 NEW FEATURES
 - add of a component in the output of RNAseqdata, continuousomicdata(),  continuousanchoringdata(), microarraydata() and (data.sd, which gives, for each item, the sd of the response per condition - NA if no replicate for a condition).
 - add a new argument in drcfit(), named deltaAICminfromnullmodel, in order to relax requirements on the information criterion to keep the best fitted model (see ? drcfit()).
+- Modification of curvesplot() to be able to put its argument dose_log_transfo by default at TRUE
+as in other functions. Curvesplot now use minimum and maximum values of the chosen BMD to 
+fix the rage on which the theoretical curve is calculated (and plotted) ad so the chosen BMD is required
+in the input of the function.
 
 BUG FIXES
 - add the sample names in column names of the output of formatdata4DRomics.